pyproject.toml
[project]
name = "ragprodapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "fpdf>=1.7.2",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.18",
    "llama-index-readers-file>=0.6.0",
    "llama-index-vector-stores-qdrant>=0.10.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
